Move more stuff from data_cloud to tests/remotes.py
